Refactor: Integrate web scraping into chat action
